splitting up table and table row attempting to find the hidden circular dep that is messing up tests.
